Improve plugin setting
